Refactor migration logic
- Split schema initialisation script out into its own file
- Represent a migration with a class to make the interface clearer
- Maintain the linear ordering of changes with an explicit list of
  migrations to apply in a specific order.
- Update the SqliteWriter to use the modified migration interface
